Struggling to figure shit out
Need to figure out how to get all types of posts in the same queryset for the index view to work properly. Not sure if Django allows this by default and I just haven't found the documentation or whether its not supported and I am doing something stupid. Will need to do more research and/or ask chatGPT