Made site work (for real this time)
Each post type now has a separate detail page, url and separate corresponding view.
Detail views now return a context object that contains the specific data for the primary key given.
Index page now creates the correct link for the content type so that that the data is passed to the correct page to display it.
Removed the old way I was trying to combine models, inheritance and chain work fine.
Added image and video admins to admin.py.