Cleaned up code & prep for authentication
Removed detail.html as it is not needed for the single detail url implementation I settled on.
Added user/<pk> and myaccount/ to urls.
Created user and myaccount views in views.py.
Formatted code.

Next to add is actual pages for user and myaccount, then authorisation. Not much data will be held on each user, just enough to prevent basic spam and allow people to post comments under the same username. It will also allow me to have a superuser account to post shit.