Bodged a cover image for blog posts
Hard coded an image as a proof of concept, but this needs to be correctly connected to the database. Until hosting properly this will be a bodge.

Updated css for index to give post containers an outline.
Updated css for blog posts to remove unneeded sections and add in cover image stuff.

Added a hosting method to image model, and methods for both image and video to get the hosting type